requirements.txt
openai-whisper
pyannote.audio==3.3.2
pyannote.core
torch
torchaudio
sounddevice
scipy
numpy
python-dotenv
